Removed unrequired write and fixed auditctl rule removal
