+++ src/monitor.rs
@@ -14,8 +14,6 @@ use std::time::{SystemTime, UNIX_EPOCH};
 use time::OffsetDateTime;
 // To use intersperse()
 use itertools::Itertools;
-// To run commands
-use std::process::Command;
 // Event handling
 use notify::event::{EventKind, AccessKind};
 
@@ -115,29 +113,10 @@ pub async fn monitor(tx: mpsc::Sender<Result<notify::Event, notify::Error>>,
     let mut last_position = 0;
     if ! config.audit.is_empty() && utils::get_os() == "linux" && utils::check_auditd() {
         for element in config.audit.clone() {
-            let mut rule: String = String::new();
+            //let mut rule: String = String::new();
             let path = element["path"].as_str().unwrap();
-            match element["rule"].as_str(){
-                Some(value) => {
-                    for c in value.chars(){
-                        match c {
-                            'r'|'R' => rule.push('r'),
-                            'w'|'W' => rule.push('w'),
-                            'a'|'A' => rule.push('a'),
-                            'x'|'X' => rule.push('x'),
-                            _ => rule = String::from("wax")
-                        }
-                    }
-                    rule.clone()
-                },
-                None => String::from("wax")
-            };
-            match Command::new("/usr/sbin/auditctl")
-                .args(["-w", path, "-k", "fim", "-p", &rule])
-                .output() {
-                Ok(d) => debug!("Auditctl command info: {:?}", d),
-                Err(e) => error!("Auditctl command error: {}", e)
-            };
+            let rule = utils::get_audit_rule_permissions(element["rule"].as_str());
+            utils::run_auditctl(&["-w", path, "-k", "fim", "-p", &rule]);
             info!("Checking audit path: {}", path);
 
             match element["allowed"].as_vec() {
@@ -168,13 +147,8 @@ pub async fn monitor(tx: mpsc::Sender<Result<notify::Event, notify::Error>>,
         match ctrlc::set_handler(move || {
             for element in &copied_config.audit {
                 let path = element["path"].as_str().unwrap();
-                match Command::new("/usr/sbin/auditctl")
-                    .args(["-W", path, "-k", "fim", "-p", "wax"])
-                    .output()
-                    {
-                        Ok(d) => debug!("Auditctl command info: {:?}", d),
-                        Err(e) => error!("Auditctl command error: {}", e)
-                    };
+                let rule = utils::get_audit_rule_permissions(element["rule"].as_str());
+                utils::run_auditctl(&["-W", path, "-k", "fim", "-p", &rule]);
             }
             std::process::exit(0);
         }) {
